stop using overpass and also only make one request
these frameworks kind of all suck.  overpass strait up throws an error if you try to load a multipolygon with its perimeter broken up into multiple ways, which I really don't think is uncommon in OSM, because it tries to pass it to some geography package that requires shapes to be nonintersecting.  brilliant idea, gies, let's try to do math that the user didn't ask for that isn't garanteed to even work for all of the data we expect to get.  so now I'm just using plain requests, and it's working great.

almost as importantly, I did some refactoring so that now it only makes one query for all of its data, and then sorts out the data into the varius types of feature before making the SVG.  I think OSM will be less mad at me this way, since they seem to not like multiple queries even if the individual queries are quite small.  now this is almost a working product; I just need to get the coastlines to consolidate correctly.