make coastlines work correctly
whew, that was way more complicated than I feel like it should have been.  I added a function to stitch together multipolygons into regular polygons and another function on top of that to add the closure to the coastline.  I had to break the coastlines and lakes into two separate layers because it turns out some water features are multipolygons whose edges are not garanteed to have any particular direction which means they don't work in the coastline closure algorithm.