change bounds order to S/N/W/E
now it's the latitudes and then the longitudes.  and if you put them in the wrong order, it will automatically fix it (this is necessary because a concerning number of Wikipedia maps have their north and south bounds swapped; probably because some genius made the bounding box template go north/south and not south/north).