construction
the way OSM handles things that are under construction is so stupid.  why is it <way highway=construction construction=primary> and not <way highway=primary consturction=yes> or something like that?  now I have to add all these if statements to check for both the keys that I'm interested in _and_ the key construction specifically when its _value_ is the key that I'm interested in.