adjust styling
I made the roads yellow, removed sand features (there's one golf course in Las Vegas with so much detail on the bunkers that trying to download the sand data crashes the query, anyway), and made the contrast slightly better.  I also added runways to the airports because otherwise it's not always clear what they are and why they're colored differently.  I also made highways even thinner when they're the only roads shown but I still think this style of map looks ugly when the scale is that small.