automatic line thickness adjustment
I realized that the main reason zoomed-out maps looked so ugly was just the dense thick lines.  I initially wanted line thicknesses to generally be constant so that someone looking at a lot of maps could get an intuitive feel for the scale, but that didn't really mae sense in retrospect.  I just don't think that really works since the spacing of the roads is not constant.  and more impotantly, it's just common sense that when you zoom out the roads get thinner.  making them stay the same if anything would create the false impression that all the maps are at the same scale.

so anyway, I generalized the thing with the thinner highways so that now all lines get thinner when there are no minor streets.  there are also now two tiers of highway, and there's less of a difference in thickness between minor streets and major streets.

I also bumped up the default street-detail by 1 because I realized I kept having to do that anyway, and the thinner lines mean this now tends to look better.